Revert "Q-IRIS related changes to expose QIREE-QSIM and QIREE-XACC APIs to IR…"
This reverts commit eb1c21ed47fad4b4242b7acb6b002bcc10085a51.
