+++ app/qir-qsim.cc
@@ -3,7 +3,7 @@
 // See the top-level COPYRIGHT file for details.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 //---------------------------------------------------------------------------//
-//! \file app/qir-qsim.cc
+//! \file qir-xacc/qir-xacc.cc
 //---------------------------------------------------------------------------//
 #include <cstdlib>
 #include <iostream>
@@ -18,8 +18,34 @@
 
 using namespace std::string_view_literals;
 
-int parse_input(int argc, char* argv[]);
+namespace qiree
+{
+namespace app
+{
 //---------------------------------------------------------------------------//
+void run(std::string const& filename, int num_shots)
+{
+    // Load the input
+    Executor execute{Module{filename}};
+
+    // Set up qsim
+    QsimQuantum sim(std::cout, 0);
+    QsimDefaultRuntime rt(std::cout, sim);
+    ResultDistribution distribution;
+
+    // Run several time = shots (default 1)
+    for (int i = 0; i < num_shots; i++)
+    {
+        execute(sim, rt);
+        distribution.accumulate(rt.result());
+    }
+
+    std::cout << distribution.to_json() << std::endl;
+}
+
+//---------------------------------------------------------------------------//
+}  // namespace app
+}  // namespace qiree
 
 //---------------------------------------------------------------------------//
 /*!
@@ -27,5 +53,22 @@ int parse_input(int argc, char* argv[]);
  */
 int main(int argc, char* argv[])
 {
-    return parse_input(argc, argv);
+    int num_shots{1};
+    std::string filename;
+
+    CLI::App app;
+
+    auto* filename_opt
+        = app.add_option("--input,-i,input", filename, "QIR input file");
+    filename_opt->required();
+
+    auto* nshot_opt
+        = app.add_option("-s,--shots", num_shots, "Number of shots");
+    nshot_opt->capture_default_str();
+
+    CLI11_PARSE(app, argc, argv);
+
+    qiree::app::run(filename, num_shots);
+
+    return EXIT_SUCCESS;
 }
